styles: simplify color parsing
Check if the color has an alpha channel before parsing, and use
gdk_color_parse() if not. Change the alpha channel parsing code to use strtoul()
instead of a homebrew solution, and add error checking to that code path.

The color parsing is now much stricter, hexadecimal colors now must begin with
'#' and may not have it at a random position in the string as before. Colors
with alpha channels must have a length of exactly 8 hex digits after the '#'. On
the other hand the additional rgb colors forms supported by gdk_color_parse()
like #rgb are now handled correctly.