simplify interface of track_clear()
It was always called with the track belonging to the appdata pointer also
passed in, so the function itself can take it from there. Afterwards that
pointer was always set to NULL, so this is now also done inside that function.
While at it disable the buttons and clear the appdata member before actually
freeing the memory for extra safety.