rework OSM data file backup handling
The old data file was renamed to backup.osm prior downloading the new file.
There was however no code to move the backup file to the original place in case
the program was interrupted (i.e. crashed) at that point. Download the new data
to a new filename instead and simply move it to the final location on success.
This way a valid data file is at the correct place at all times (given that one
existed before), and the worst case scenario is some wasted space until the next
project download.

While at it actually check the error code returned by net_io_download_file()
instead of only looking if a file with the desired name was created. This is
what net_io_download_file() already ensures, so just do this for some extra
sanity.