shrink sizeof(base_object_t) on 32 bit platforms
Reorder the fields and reduce the size of both visible and flags so the whole
struct fits into 32 byte. This are 4 byte saved for the overall size as well
as 4 bytes padding. Since this struct is embedded into the structs of all OSM
objects this should reduce the memory usage for all objects in a project.