introduce and use track_is_empty()
The users of track_seg_points() in map.c did not really care about the number of
elements in the track, they only wanted to see if it was empty. One even did not
use the information at all. The only user is now inside track.c, so mark it
static.

Introduce track_is_empty() which abstracts that check by just looking at the
pointer. This avoids iterating the entire list, which is not only needless
computation, is also trashes the cache, which causes further slowdown.