add a submenu for all used presets of an object
If an item is selected that matches all static keys of a given interactive
preset, then a submenu is inserted at the top of the presets list which contains
references to all used presets. In case only certain keys of this preset are to
be edited this allows a much faster access to the templates.

This is currently only for the desktop version.