improve track_read() usage
Until now track_read() always set the track to dirty, which one of 2 callers
reset, and the other did it again. Also it calls track_info(), which one of the
callers did again.

Only call track_info() in track_read(), and pass the desired dirty state right
to the function.