fix memleak of dash styles
When setting the line-dash object property another reference is taken to the
dash object, in addition to the one every object has on creation. Do not take
another one in this function, otherwise the reference count will never drop to
zero again. Instead lower the reference count once the property has been set, so
the dash style will be freed when the way objects drops it's reference on it
(e.g. because of its own destruction).