avoid most memory allocations and many syscalls in file_scan()
Check if the filename has the given extension first, before allocating the full filename or
checking if it is a regular file. Since the extension check is a pure userspace operation
this avoids a context switch to kernel and back, also the extension check on the shorter
string is faster.