reorganize tag add and edit code
Until now for a new tag an entry in the tag model was created, selected, and the usual tag edit code
was called. Now the function that handles the tag edit operates on a passed tag_t* without caring for
the model or anything. Only the actual tag edit code now scans the model for the selected row, the tag
add code will only touch the model if a new tag was actually inserted. While at it this now prevents
entirely empty tags from being inserted.