remove elemstyles_filename member from style_t
This is in a union with the actual style, which makes it easy to accidentially
break it when refactoring (which is what I initially did here). The filename is
only needed at one place, so let it be requested out of the parsing function
when needed and ignore it otherwise.