avoid iterating the whole node chain just for a message
This code is only entered if the node is either the first or last of a way.
Since it has already been checked that it is not the first there is no need
to iterate the whole chain and check if it is the last. Even worse that this
is done a few lines later anyway to find the place to insert.