speed up applying styles to nodes
When it is already clear that a rule does not match stop processing the further conditions, as the
final outcome can never be a match again. The most obvious case is a node without tags, which is
not explicitely handled here as the osm_*() functions in both branches will immediately return in
such case.

This hasn't originally been a problem when only a single condition was allowed per rule, but since
multiple conditions per rule are now supported this needs to be handled to avoid slowdown.