avoid writing the whole track XML when saving
The common case is that the XML file has been written by OSM2go before, so the
structure is known and well defined. Skip to the last <trkseg> entry in the file
and start writing here, keeping all previous ones unchanged. This especially
saves the need to call e.g. localtime_r() for all timestamps. In case of any
unexpected things happening just fall back to the old behavior and write the
complete file.