remove JOSM_PATH_ADJUST
This was never defined, since JOSM splits it's images between styles and
presets. If it would ever be defined a dangerous situation would arise as memory
managed by GTK functions and those managed through libxml2 functions would be
mixed, which may lead to nasty effects if any of these ever do custom code and
not just pass on calls to libc functions.