free unused memory for track vectors
std::vector allocates storage in bigger chunks, for practical cases it just
doubles it's size once the internal array overflows. A track segment that was
loaded from disk will never be appended to, so copy the vector to one of the
exactly right size when the segment ends. This will cause some additional
memory allocations to happen, but the memory usage once the track was loaded is
greatly reduced.