shrink and inline elemstyle_line_mod_t into elemstyle_t
This has now the same size as the former pointer on 32 bit, and only half the
size on 64 bit systems. This avoids one needless memory indirection per
"linemod".

The reduced size for width doesn't matter, this is a pixel size on screen, having
anything even close to 100px isn't useful anyway.