improve updating GPS track drawing
The previous code was very much derived from the old C code that based on a
single linked list. This is already using a vector, so finding the last and
second last entries is much simpler using iterators. Only 2 checks are needed
now to find out if anything has to be drawn at all. Searching for the first
point to draw can now start at the end and only needs to run until the first
invisible item is found, so anything before that in the track segment will not
be checked at all anymore.

This should considerably reduce the load, especially when the current position
is outside the project bounds.