fix list of all relations always empty
This removes some dead code on the way. This was broken by the combination of
commit 5337dd828108a9d7a30f893cea217b335d0d8e4f, which removed the explicit
assignment of NULL to this member, and commit
fe4ea4040a2a799d7b42a2be10d7c71d46099946, which removed the general
initialization of all members to 0. Remove the member "object" and all code
checking it as it was never needed.