always increment row when something was attached to presets widget table
Separators add something to the table but don't return a widget as they are no
inputs. This caused the next widgets to be in the same table row, leading to
visual disortion. Pass the row counter by reference and let the underlying
functions that actually modify the table do the incrementing.