use stag_t for intermediate storage during tag edit
Using a tag_t without allocating the tag memory can't work: it can't be
distinguished if the tag was newly created during edit and the memory needs to
be freed or if the tag was present before and the memory needs to be kept.
Simply use stag_t when the memory needs to be allocated anyway as it makes
handling easier.