rework user id handling
Also but users without id into the normal map, just use negative ids for them
that are generated when the username is first observed.

This also avoids possible stale pointer accesses when the internal data of the
std::string is stored in the objects, but the strings can be moved around if the
map reallocates and moves the contents.

Additionally it saves 4 byte per OSM object on 64 bit platforms.