reduce appdata_t usage in style handling
The main objective is to not need settings to be loaded when calling
style_load() in the style_apply testcase. Calling settings_load() even when
completely ignoring the return code breaks the test for unknown reasons. Running
the test under valgrind make the error go away.

Since using less appdata_t internals is a good thing to do anyway regroup the
code to only get the actually required values.