rework track_pos2lpos()
Half of the callers did not actually care for the return value, make them call
pos2lpos() directly. Other callers were only interested in the fact, if the
point is visible on the screen. To determine this no calculation is needed, just
compare to the lat/lon values of the project bounds instead.