split diff_restore_file() out of diff_restore()
The latter takes just an appdata_t reference, as this is what all usual callers
pass in. The new function does not use any of appdata_t internals, and it
returns a combination of flags to denote what exactly the outcome was. This is
used in the tests only for the moment.

While at it do some small improvements to the error message like marking them
translatable.