remove osm_t::to_relation()
This iterates all relations, find the place of a given member or it's absence,
and puts the relations into a vector. Later it iterates through all matched
relations to find the member again. Simply do the processing in one step,
eliminating the duplicate iteration of the matching relations as well as the
allocations for the intermediate vector.