optimize render_arrow() for project map editing
The sin() and cos() calculations are actually constants, but at least the N900
compiler was not able to figure that out so it did that calculations at runtime,
also needlessly promoting the precision to double.

With this table of constants the other arguments to render_arrow() can also be
expressed in terms of this table, simplifying the function interface. At the end
the compiler is now able to inline the function into the caller.