get rid of the detail popup
There is enough room in the desktop version for the 2 detail buttons. The detail
popup has problems, e.g. when the popup is opened and then the switching virtual
desktops, which keeps the popup visible also on the other desktop. Instead of
putting effort into this simply get rid of it.