delay map autosave enabling on startup
It doesn't make sense to start the timer before a project is loaded. The current
implementation would enable it in the map_t constructor, which is called from
the derived map_internal constructor before the members of the child class are
set up, so it may even operate on uninitialized data.