rework presets_context_t handling
There can only be one instance of this at any given time, so create a static
instance variable in the class. This allows to omit the context from several
callbacks, which in turn allows to pass more interesting values to the callbacks
and avoid to store either the context or the additional value as GObject
property.

Remove the pointer from the preset_attach_context struct. Only one class did
actually use it, and only to request the icon singleton. If it would be needed
one could easily embed that pointer into the preset_attach_context struct that
is passed anyway.

Make the functions that construct the dynamic presets submenus members of
presets_context_t.