optimize joining objects
When joining nodes or ways the list of relations was iterated at least 2 times:
once for each of the original objects to see if it was member in any relation.
If the object that was slated for removal was member of any relations the list
of relations was iterated a third time to do the replacement.

Now the list is operated only once, saving the affected relations into vectors.
The vector of the object slated for removal is kept and later iterated through.