get rid of osm2go_platform::Pixmap
The main point is to let canvas_t::image_new() directly deal with icon_items.
The code to convert that back to a pixmap is now in the platform specific code,
so other platforms are free to use entirely different things.

The idea behind that is to allow the canvas_item to make better use of SVG icons
instead of rasterizing them right at the start, which destroys every possible
visual improvement during scaling.