cleanup canvas_item_t destroy notification
Now we have a pure virtual destroy base class. This one does the delete things
in a virtual method, which is just one pointer indirection for the call as
before with the handler pointer. It has the advantage that the handler is now
not stored again for every object, but there is only one virtual table per
type.

Also the canvas_item_t can be removed from canvas_item_info_t, it was only
needed to run exactly this cleanup. It is however present at the point when
glib notifies that the object is about to be removed, so just pass on the
pointer there.

Also move canvas from the item to the destroyer. This makes the type member
the first and only member in canvas_item_info_t, so it always is in the first
cacheline when accessing these object. One more member of a derived object also
fits in there. The canvas member is only needed for the cleanup, so it doesn't
matter where it's stored and can just sit in the destroyer object.