put map_highlight directly into map_t
The indirection only helps to keep the internals out of map.h, which is only a
compile time thing. On the other hand the current construct needs one additional
pointer for the reference, and the destructor was not run because the destructor
is not virtual, so a bit of memory was leaked on program termination. This isn't
critical, but since there are no real benefits just do it properly.