do bounded scrolls also in canvas_t::scroll_step()
This avoids the extra step call to canvas_t::scroll_to() with the result of this
call. It especially avoids conversions between floats and ints that lead to
scrolling in both directions at the same time when either being close to the
border or in high zoom levels because of the precision loss that happens during
these conversions.