prevent that duplicate tags are entered at all
The user may now not create a new tag collision either by adding a tag or by
modifying an existing one. Collisions may now happen only when e.g. merging
objects or loading diffs with collisions. This simplifies the code to insert
or modify tags as the only collision change that may happen is the removal of an
already existing one.