fix uninitialized variable being used for object version after initial upload
The second check to obj->isNew() will always return false, after an upload every
object has a positive id. Remove this scary microoptimization.

This does not cause any harm as this is only used to move the new objects out of
that state, which has always worked. The OSM data is downloaded and loaded again
afterwards, so the bad version value should usually not have been visible to
anyone.