do not reserve storage for a second map_item in every map_item_chain_t
An item can have multiple map items. This is only used for ways that have a
background painted, which are only a fraction of the actual drawn elements.
Those additional background items are now stored in map_t, saving one pointer
for the majority of visible_object instances that do not need a background item.