do not save in a map item if it is a highlight item or not
The only place where this was needed is to find the primary map object for a
clicked item. Since both the highlight items and the normal ones refer to the
OSM object, and that has a pointer of it's primary map object simply jump the
chain once up and down as it already was done for highlight objects. Doing it
for real ones must work as well. This saves 4 byte per visible object, 8 byte on
64 bit platforms because of padding.