further optimize tag operations when reversing ways
A strcasecmp() with a fixed string is much more likely to get optimized by the
compiler. Since the string compared to are short (3 and 4 byte) the whole memory
allocation for the lowercase std::string can be entirely avoided.

This introduces tag_t::value_compare() and tag_t::key_compare() to make it
obvious that these are intentionally only comparing the pointers as both sides
should come from the value cache.