simplify checks for tag collision updates
Use std::distance() to calculate the amount of tags with the old key that have
been there before. This is slightly less efficient if there are more than 2
instances of the old key, but this is very unlikely. Even if there are many
more then iterating is not very expensive anyway. On the other hand this allows
to skip updating the collision tag if one of the colliding keys has been changed
but there are still 2 or more left.