rework opening a project from the selection list
Until now the name of the project was returned, but the whole project_t object
that already had the project file parsed and various other things was discarded.
Afterwards a new project_t was created just from the given name. Use the
obvious shortcut and directly create a new project_t in project_select() and
move all information already known over to this.