mark also the second variant of josm_icon_name_adjust() nonnull
There were only 2 callers: one is the other variant, which is already nonnull,
and a callback from the XML SAX parser. Looking at the libxml2 code it already
checks that both attribute key and value are not NULL before processing
further, and this has been the case for all releases from this millenium.