fix swapping with trstring::toStdString()
This works fine in the Gtk version as trstring actually is a std::string. In
case toStdString() only returns a std::string, e.g. when trstring is QString,
this would pass a reference to a temporary into swap(). Reverse the object order
so trstring can provide a matching swap() implementation so this always works.
Remove the non-const version of trstring::toStdString() to avoid further
problems with this difference.