style_apply test: fix reference to already deleted pointer
The oldicon variable holds the address of the previous item before the style
changed. It compares only the pointer, so things are safe. But when another
style change happens it is perfectly legal that a new item will end up on the
same address, so the test may fail. Update the stored value in between as the
thing that should be checked here is that the _last_ style change did cause an
icon change.