rework map_state_t
The idea of moving this into one central place was that map_t can change it when
the view is changed (i.e. moved), and the active project can easily save that
to storage. That worked, but then I removed the code that completely reloaded a
project when selected in the project list. There was just one problem: all
projects in that list shared a single dummy map_state_t, so the actual
information from the project file was already lost.

Do it the other way: put the map_state_t in project_t, and just update it in the
currently open project.