fix disappearing track segments when segment started offscreen and is appended
When a track segment starts offscreen the last offscreen point is still used in
drawing so the first drawn line enters the screen. This was however not
accounted for so when the first point was appended to the segment then this
initial point was not drawn anymore on the next step, so the track segment
suddently started on the screen.