osm-gps-map: make tile cache purging less aggressive
Do not purge all tiles that do not match the current redraw cycle, but only
about half of them. This makes it more likely that when zooming back and forth
the old tiles still are in the cache. Otherwise the cache will simply be purged
again.

While at it remove a needless int/pointer indirection when passing the purge
limit to the test function.