osm-gps-map: use an integer to hash tiles
Most parts of the URL are constant for a given map provider (and only OSM is
currently supported anyway). The only variable parts of the URL are the tile
coordinates. Use an uint64 to store these values and hash them instead of the
string, which is much more efficient. Also it is not necessary anymore to keep
the "filename" around all the time. It is not used as filename anyway as the
tiles are only kept in memory, so it was only used to hash the input values.