osm-gps-map: invert the custom log handling logic
The custom log handler was defined for all messages, and passed all non-debug
ones to the default handler. If ENABLE_DEBUG was set it passed those, too.
Register the handler only for the debug messages at all, so all others will
directly go to the default handler. If ENABLE_DEBUG is set don't do anything.