osm-gps-map: use std::pair of coordinates for tracks and bounds
Since only rectangles are used anyway this reduces the code and memory size.
Also switch to std::vector for storage of bounds, as it also simplifies handling
a lot and makes it more type safe.

An additional benefit is that the redraw area for new rectangles is not always
starting at (0, 0), but only at the real minimum coordinates of the rectangle.