avoid a memory allocation in trstring::arg(const char *)
While at it fix wrong replacement of %n in constructor if the translated text
does not have the placeholder at the exact same position as the original string.
This also removes the arbitrary limit to the number of supported placeholders
and just supports them up to %9 as that is just the same code.