improve logic to enable "open project" button
The dialog is already enabled for all existing projects, so this code path only
relevant for new ones. Besides being likely impossible to hit all branches that
do not result in a somehow valid file being present do not make sense to end up
enabling the dialog.

Fixing the logic was easy, but it took me multiple attempts to understand why
the code should really be this way. I guess this is a good sign that the old
code was too complicated.