optimize osm_t::node_delete() for nodes not part of any way
In this case iterating through all ways can simply be skipped. There was only
one place that expected a non-empty return value in that case, and that one only
asserted that it was the correct way returned. Since this has never failed just
remove that code and skip the iteration.