make sure the test objects are consistent regarding their basic properties
When creating a base_object_t the flags were set to dirty for new objects, i.e.
such objects that were not yet uploaded. These should have a negative id and a
version of 0. When calling osm_t::attach_(node,way,relation)() a new id was
searched and the existing id in the object was overwritten. Make sure that
there is always version 0 and ID_ILLEGAL for objects passed through these
methods.