get rid of way_t::cleanup()
It was called from 2 places, which actually come from osm_t::way_delete(), were
the node chain is actually cleared before reaching that point. The 2 other
callers of osm_t::way_free() are the upload code, where the way is internally
marked as deleted before and has the node chain already cleared, and a misuse in
the testcases. The clearing of the tags is actually not needed at all as either
the object is deleted, which would delete the tags as well, or the tag list is
already cleared by markDeleted().