fix possible relation roles not shown for multipolygons
If a multipolygon matches another preset earlier that one usually does not have
roles because it uses the multipolygon as "just another kind of way". In that
case the roles of the multipolygon were not applied.

There still could be cases where one relation matches multiple presets, but at
the moment all relations set a "type" value and possible other tags so there
should be only one match. The solution would be to store multiple items, but
let's not do this until there is actually use for the code.