microoptimize some math operations when finding way segments
Do not use pow() to calculate i^2 for an integer as double, and then truncate
to a float. Use a direct multiplication, this is as readable and may even be
faster if the compiler doesn't optimize the call away anyway. Still do it as
floats to avoid possible integer overflows for very long way segments.

Afterwards some casts can be removed as the result would be a float anyway
because the other partner of the operation now is a float.