refactor relation_t::descriptive_name()
This allows osm_t::unspecified_name() to better handle the result, e.g. there
is no need to check if the first character is '<' to see if there has been a
proper name tag found of if this is an id. In fact there are now even less
memory allocations in case there are multiple relations containing this object
as the conversion to std::string is only done when the name is actually better
than the previous one.