tests: fix item selection test
The new item was just at the border of the rectangle that may intersect with the
provided items: the right edge of the polygon was at FUZZ_PIXELS*3, the click
position was at FUZZ_PIXELS*4, so looking FUZZ_PIXELS in every direction may
actually return the item. Add one extra pixel to ensure this doesn't happen.

That the code does actually _not_ intersect this properly is a different bug.