make style_t an abstract base class for styles
There currently is only the josm_elemstyles implementations, but the relaying
from style.h to josm_elemstyle.h functions can as well be just be done by a
virtual function. This allows also a bit more guarding in the testcases.

As a positive side effect the case where parsing the elemstyles file fails can
now be properly be detected and the style is not returned to the caller.
Previously a style without information was returned that would just do nothing.