fix Clang warnings about defaulted operator
Clang does not allow that the defaulted assignment operator overwrites const
members. Actually this is not needed at all for C++11, where only a copy
constructor is needed, and that can safely be defaulted. The only exception is
gcc 4.2, that requires the assignment operator to make std::vector::push_back()
to work.