check for "building:part" before checking if the object has only a single tag
If this is the only tag it would have been found anyway, and if not the
outcome would be basically the same. The difference now is that even for an
object that only is tagged as "building:part" the beautified name with a
space is used instead of the raw key.