show info dialog after adding a way only if the new way has no tags
If an old way was extended or the way was joined with a different way ending
it may be that there is a valid set of tags already, no need to show the info
dialog then.

This has the nice advantage that one can cheat around the info dialog during
the testcase, so a way can actually be added through the "UI" during the tests
now.