drop new nodes when a copy already exists in the data
When an upload gets interrupted or similar things happen one may end up in a
situation where the diff still contains new nodes that were in fact successfully
uploaded. For every new node present when loading a diff scan the database if an
identical copy already exists, and simply drop the new node in this case. The
references in modified ways and relations need to be fixed up as well.