string_view: provide nonstd::string_view::ends_with() even for C++17
When the compiler supports C++17 it's std::string_view would be used for
nonstd::string_view, which usually does not contain the ends_with() method. This
leads to the strange situation where both older and newer compilers support the
ends_with() API, but not the versions in between. Do not simply forward the
class, but create a subclass that implements the missing API as well.