fix overridden server not being present after project was opened
When a project was saved with the correct server setting and then opened on
project start the correct server was used. But when opening the project from the
project list the correct server was loaded to the list of projects, but
afterwards not transferred to the actually opened project, so the default server
was always used.

This resulted in creation of https://www.openstreetmap.org/changeset/139316267
by accident.