map: avoid assertion on bogus click events
I can reproduce a failed assertion because of the wrong object type roughly as
this on the desktop version:

-select a node
-double click on it
-open the properties window
-drag the window so it is not above the node
-double click on any property line in the window

This is likely a bug in the Gtk click handling, not necessarily in Gtk but more
likely just in the related code of OSM2go itself. It seems like the click event
would not be consumed on the double click in the properties window.